crud

+++ src/books/services.ts
@@ -1,4 +1,6 @@
 import {
+  BadRequestException,
+  ForbiddenException,
   Injectable,
   NotFoundException,
   UnauthorizedException,
@@ -45,26 +47,53 @@ export class BookServices {
     return createdBook;
   }
 
-  async deleteBook(id: string) {
-    const { data, error } = await this.supabase
+  async deleteBook(id: string, userId: string) {
+    const { data: bookToDelete } = await this.supabase
+      .from('books')
+      .select()
+      .eq('id', id)
+      .single();
+    if (!bookToDelete) {
+      throw new NotFoundException('Book with this ID is not found.');
+    }
+    if (bookToDelete.user_id !== userId) {
+      throw new ForbiddenException('Forbidden action for this user.');
+    }
+
+    const { error } = await this.supabase
       .from('books')
       .delete()
       .eq('id', id)
       .select();
-
-    if (!data || data.length === 0) {
-      throw new NotFoundException();
+    if (error) {
+      throw new BadRequestException(error);
     }
   }
 
-  async updateBook(id: string, body: any) {
-    const { data } = await this.supabase
+  async updateBook(id: string, body: any, userId: string) {
+    const { data: bookToUpdate } = await this.supabase
+      .from('books')
+      .select()
+      .eq('id', id)
+      .single();
+    if (!bookToUpdate) {
+      throw new NotFoundException('Book with this ID is not found.');
+    }
+    if (bookToUpdate.user_id !== userId) {
+      throw new ForbiddenException('Forbidden action for this user.');
+    }
+    
+    const { data, error } = await this.supabase
       .from('books')
       .update(body)
       .eq('id', id)
-      .select();
-    if (!data || data.length === 0) {
-      throw new NotFoundException();
+      .select()
+      .single();
+    if (error) {
+      throw new BadRequestException(error.message);
+    }
+    if (!data) {
+      throw new NotFoundException('Book with this ID is not found.');
     }
     return data;
   }
